requirements.txt
Flask>=3.1.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
Jinja2>=3.1.0,<4.0.0
MarkupSafe>=2.1.0,<3.0.0
itsdangerous>=2.1.0,<3.0.0
click>=8.1.0,<9.0.0
blinker>=1.9.0,<2.0.0
gunicorn>=21.2.0,<22.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
Pillow==11.3.0